Move inference modules to `inference` package
